\name{Swisssfm2-package}
\alias{Swisssfm2-package}
\alias{Swisssfm2}
\docType{package}
\title{
\packageTitle{Swisssfm2}
}
\description{
\packageDescription{Swisssfm2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Swisssfm2}
\packageIndices{Swisssfm2}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Swisssfm2}

Maintainer: \packageMaintainer{Swisssfm2}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
