\name{Swisssfm-package}
\alias{Swisssfm-package}
\alias{Swisssfm}
\docType{package}
\title{
\packageTitle{Swisssfm}
}
\description{
\packageDescription{Swisssfm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Swisssfm}
\packageIndices{Swisssfm}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Swisssfm}

Maintainer: \packageMaintainer{Swisssfm}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
